Move all full_text_search parts to server-kotlin module
