--- fulltext-search/build.gradle
@@ -1,48 +1,0 @@
-buildscript {
-    repositories {
-        jcenter()
-    }
-    
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-group 'cz.cityvizor'
-version '0.1.0'
-mainClassName = "io.ktor.server.netty.EngineMain"
-
-sourceSets {
-    main.kotlin.srcDirs = main.java.srcDirs = ['src']
-    test.kotlin.srcDirs = test.java.srcDirs = ['test']
-    main.resources.srcDirs = ['resources']
-    test.resources.srcDirs = ['testresources']
-}
-
-repositories {
-    mavenLocal()
-    jcenter()
-    maven { url 'https://kotlin.bintray.com/ktor' }
-}
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    compile "io.ktor:ktor-server-netty:$ktor_version"
-    compile "ch.qos.logback:logback-classic:$logback_version"
-    compile "io.ktor:ktor-server-core:$ktor_version"
-    compile "io.ktor:ktor-jackson:$ktor_version"
-    compile "io.ktor:ktor-client-core:$ktor_version"
-    compile "io.ktor:ktor-client-core-jvm:$ktor_version"
-    compile "io.ktor:ktor-client-jetty:$ktor_version"
-    compile group: 'org.apache.lucene', name: 'lucene-core', version: '8.3.0'
-    compile group: 'org.apache.lucene', name: 'lucene-queryparser', version: '8.3.0'
-    compile 'org.jetbrains.exposed:exposed:0.17.7'
-    compile("org.postgresql:postgresql:42.2.2")
-    compile 'com.fasterxml.jackson.datatype:jackson-datatype-joda:2.7.3'
-    compile group: 'org.apache.lucene', name: 'lucene-analyzers-common', version: '8.3.0'
-    implementation 'com.github.doyaaaaaken:kotlin-csv-jvm:0.7.3'
-    testCompile "io.ktor:ktor-server-tests:$ktor_version"
-}
