--- fulltext-search/gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,0 @@
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.10-all.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
