--- fulltext-search/resources/application.conf
@@ -1,14 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-        port = ${?PORT}
-    }
-    application {
-        modules = [ cz.cityvizor.ApplicationKt.module ]
-    }
-    cityvizor {
-        jdbcUrl = "jdbc:postgresql://localhost:5432/test"
-        dbUser = "test"
-        dbPass = "test"
-    }
-}
