--- fulltext-search/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = "fulltext-search"
