--- fulltext-search/test/ApplicationTest.kt
@@ -1,26 +1,0 @@
-package cz.cityvizor
-
-import io.ktor.application.*
-import io.ktor.response.*
-import io.ktor.request.*
-import io.ktor.routing.*
-import io.ktor.http.*
-import com.fasterxml.jackson.databind.*
-import io.ktor.jackson.*
-import io.ktor.features.*
-import io.ktor.client.*
-import io.ktor.client.engine.jetty.*
-import kotlin.test.*
-import io.ktor.server.testing.*
-
-class ApplicationTest {
-    @Test
-    fun testRoot() {
-        withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Get, "/").apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertEquals("HELLO WORLD!", response.content)
-            }
-        }
-    }
-}
