+++ server-kotlin/resources/application.conf
@@ -6,6 +6,11 @@ ktor {
     application {
         modules = [ main.ApplicationKt.module ]
     }
+    cityvizor {
+        jdbcUrl = "jdbc:postgresql://localhost:5432/cityvizor"
+        dbUser = "postgres"
+        dbPass = "pass"
+    }
 }
 
 akka {
