+++ server-kotlin/src/full_text_search/SearchRouter.kt
@@ -1,15 +1,17 @@
-package cz.cityvizor.router
+package full_text_search
 
-import cz.cityvizor.model.FTSException
-import cz.cityvizor.service.combineLucineAndDBData
-import cz.cityvizor.service.search
+import full_text_search.model.FTSException
+import full_text_search.service.combineLucineAndDBData
+import full_text_search.service.search
 import io.ktor.application.call
 import io.ktor.http.HttpStatusCode
 import io.ktor.response.respond
 import io.ktor.routing.Routing
 import io.ktor.routing.get
+import org.apache.lucene.store.RAMDirectory
 
-fun Routing.searchRouter() {
+val memoryIndex = RAMDirectory()
+fun Routing.fullTextSearchRouter() {
     get("/search") {
         val query = call.request.queryParameters["query"] ?: throw FTSException("No query supplied")
         val profile = call.request.queryParameters["profile"]
