+++ server-kotlin/src/full_text_search/model/Models.kt
@@ -1,4 +1,4 @@
-package cz.cityvizor.model
+package full_text_search.model
 
 import org.joda.time.LocalDate
