+++ server-kotlin/src/full_text_search/service/MappingService.kt
@@ -1,22 +1,22 @@
-package cz.cityvizor.service
+package full_text_search.service
 
-import cz.cityvizor.model.InvoiceLucineModel
-import cz.cityvizor.model.InvoiceResultModel
-import cz.cityvizor.model.Invoices
-import cz.cityvizor.model.Invoices.amount
-import cz.cityvizor.model.Invoices.counterpartyId
-import cz.cityvizor.model.Invoices.counterpartyName
-import cz.cityvizor.model.Invoices.date
-import cz.cityvizor.model.Invoices.description
-import cz.cityvizor.model.Invoices.event
-import cz.cityvizor.model.Invoices.eventSrcId
-import cz.cityvizor.model.Invoices.id
-import cz.cityvizor.model.Invoices.item
-import cz.cityvizor.model.Invoices.paragraph
-import cz.cityvizor.model.Invoices.profile
-import cz.cityvizor.model.Invoices.type
-import cz.cityvizor.model.Invoices.year
-import cz.cityvizor.model.ResultModel
+import full_text_search.model.InvoiceLucineModel
+import full_text_search.model.InvoiceResultModel
+import full_text_search.model.Invoices
+import full_text_search.model.Invoices.amount
+import full_text_search.model.Invoices.counterpartyId
+import full_text_search.model.Invoices.counterpartyName
+import full_text_search.model.Invoices.date
+import full_text_search.model.Invoices.description
+import full_text_search.model.Invoices.event
+import full_text_search.model.Invoices.eventSrcId
+import full_text_search.model.Invoices.id
+import full_text_search.model.Invoices.item
+import full_text_search.model.Invoices.paragraph
+import full_text_search.model.Invoices.profile
+import full_text_search.model.Invoices.type
+import full_text_search.model.Invoices.year
+import full_text_search.model.ResultModel
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.transactions.transaction
